+++ rasa_core/trackers.py
@@ -15,14 +15,16 @@ from typing import List
 
 from rasa_core import utils
 from rasa_core import events
-from rasa_core.conversation import Dialogue, Topic
-from rasa_core.events import UserUttered, TopicSet, ActionExecuted, \
-    Event, SlotSet, Restarted, ActionReverted, UserUtteranceReverted, BotUttered
+from rasa_core.conversation import Dialogue
+from rasa_core.events import UserUttered, ActionExecuted, \
+    Event, SlotSet, Restarted, ActionReverted, UserUtteranceReverted, \
+    BotUttered, TopicSet
 
 logger = logging.getLogger(__name__)
 
 if typing.TYPE_CHECKING:
     from rasa_core.actions import Action
+    from rasa_core.domain import Domain
 
 
 class DialogueStateTracker(object):
@@ -37,15 +39,12 @@ class DialogueStateTracker(object):
         the tracker, these events will be replayed to recreate the state."""
 
         evts = events.deserialise_events(dump_as_dict)
-        tracker = cls(sender_id, domain.slots, domain.topics,
-                      domain.default_topic)
+        tracker = cls(sender_id, domain.slots)
         for e in evts:
             tracker.update(e)
         return tracker
 
     def __init__(self, sender_id, slots,
-                 topics=None,
-                 default_topic=None,
                  max_event_history=None):
         """Initialize the tracker.
 
@@ -59,10 +58,6 @@ class DialogueStateTracker(object):
         self.events = self._create_events([])
         # id of the source of the messages
         self.sender_id = sender_id
-        # available topics in the domain
-        self.topics = topics if topics is not None else []
-        # default topic of the domain
-        self.default_topic = default_topic
         # slots that can be filled in this domain
         self.slots = {slot.name: copy.deepcopy(slot) for slot in slots}
 
@@ -75,8 +70,6 @@ class DialogueStateTracker(object):
         self._paused = None
         # A deterministically scheduled action to be executed next
         self.follow_up_action = None
-        # topic tracking
-        self._topic_stack = None
         self.latest_action_name = None
         self.latest_message = None
         self.latest_bot_utterance = None
@@ -112,6 +105,13 @@ class DialogueStateTracker(object):
             "paused": self.is_paused(),
             "events": evts
         }
+
+    def past_states(self, domain):
+        # type: (Domain) -> deque
+        """Generate the past states of this tracker based on the history."""
+
+        generated_states = domain.states_for_tracker_history(self)
+        return deque((frozenset(s.items()) for s in generated_states))
 
     def current_slot_values(self):
         # type: () -> Dict[Text, Any]
@@ -162,23 +162,6 @@ class DialogueStateTracker(object):
         """Return a list of events after the most recent restart."""
         return list(self.events)[self.idx_after_latest_restart():]
 
-    @property
-    def previous_topic(self):
-        # type: () -> Optional[Text]
-        """Retrieves the topic that was set before the current one."""
-
-        for event in reversed(self.events_after_latest_restart()):
-            if isinstance(event, TopicSet):
-                return event.topic
-        return None
-
-    @property
-    def topic(self):
-        # type: () -> Topic
-        """Retrieves current topic, or default if no topic has been set yet."""
-
-        return self._topic_stack.top
-
     def init_copy(self):
         # type: () -> DialogueStateTracker
         """Creates a new state tracker with the same initial values."""
@@ -186,8 +169,7 @@ class DialogueStateTracker(object):
 
         return DialogueStateTracker(UserMessage.DEFAULT_SENDER_ID,
                                     self.slots.values(),
-                                    self.topics,
-                                    self.default_topic)
+                                    self._max_event_history)
 
     def generate_all_prior_trackers(self):
         # type: () -> Generator[DialogueStateTracker, None, None]
@@ -229,6 +211,9 @@ class DialogueStateTracker(object):
                 # listen action).
                 undo_till_previous(UserUttered, applied_events)
                 undo_till_previous(ActionExecuted, applied_events)
+            elif isinstance(event, TopicSet):
+                logger.warn("Topics are deprecated, therefore the TopicSet "
+                            "event will be ignored")
             else:
                 applied_events.append(event)
         return applied_events
@@ -331,8 +316,6 @@ class DialogueStateTracker(object):
         self.latest_message = UserUttered.empty()
         self.latest_bot_utterance = BotUttered.empty()
         self.follow_up_action = None
-        self._topic_stack = utils.TopicStack(self.topics, [],
-                                             self.default_topic)
 
     def _reset_slots(self):
         # type: () -> None
@@ -360,7 +343,7 @@ class DialogueStateTracker(object):
         return deque(evts, self._max_event_history)
 
     def __eq__(self, other):
-        if isinstance(other, type(self)):
+        if isinstance(self, type(other)):
             return (other.events == self.events and
                     self.sender_id == other.sender_id)
         else: