+++ tests/test_dsl.py
@@ -154,8 +154,12 @@ def test_visualize_training_data_graph(tmpdir, default_domain):
 
     # we can't check the exact topology - but this should be enough to ensure
     # the visualisation created a sane graph
-    assert set(G.nodes()) == set(range(-1, 14))
-    assert len(G.edges()) == 16
+    assert (set(G.nodes()) == set(range(-1, 13)) or
+            set(G.nodes()) == set(range(-1, 14)))
+    if set(G.nodes()) == set(range(-1, 13)):
+        assert len(G.edges()) == 14
+    elif set(G.nodes()) == set(range(-1, 14)):
+        assert len(G.edges()) == 16
 
 
 def test_load_multi_file_training_data(default_domain):