+++ tests/test_tracker_stores.py
@@ -8,7 +8,7 @@ from rasa_core.domain import TemplateDomain
 from rasa_core.events import SlotSet, ActionExecuted, Restarted
 from rasa_core.tracker_store import InMemoryTrackerStore
 
-domain = TemplateDomain.load("data/test_domains/default_with_topic.yml")
+domain = TemplateDomain.load("data/test_domains/default.yml")
 
 
 def test_get_or_create():